+++ src/stdlib/control.rs
@@ -14,18 +14,3 @@ pub(crate) fn else_(vm: &mut VM, _: ()) -> RantStdResult {
   vm.resolver_mut().attrs_mut().make_else();
   Ok(())
 }
-
-pub(crate) fn break_(vm: &mut VM, val: Option<RantValue>) -> RantStdResult {
-  vm.interrupt_repeater(val, false)?;
-  Ok(())
-}
-
-pub(crate) fn continue_(vm: &mut VM, val: Option<RantValue>) -> RantStdResult {
-  vm.interrupt_repeater(val, true)?;
-  Ok(())
-}
-
-pub(crate) fn return_(vm: &mut VM, val: Option<RantValue>) -> RantStdResult {
-  vm.func_return(val)?;
-  Ok(())
-}